requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
requests==2.24.0
python-dotenv==0.14.0
gunicorn==20.0.4